Cache loaded configuration data by game.
The primary mechanism for loading the game data is shifted to the games
module. This module knows where the configuration data is stored for each
game (whether in this app or routes18xx), and will automatically retrieve
it based on which game we're in. It will also cache that data based on
game, so it can always be retrieved quickly.

Some of the configuration it loads is actually game objects from the
routes18xx library, such as the board. Again, it automatically detects
the game (via g.game_name) and produces the correct object, through the
routes18xx Game object.