Make the app a Blueprint to capture the game name.
Each game's app with be "/games/<game_name>". The Blueprint is setup to
strip the game name and store it in g, then inject that value into any
url_for calls. This provides an elegant and simple way for every
controller to know what game this is if it needs to.