Generalize Chicago station placement to all split cities.
Split city stations will be indicated by a single entry in the stations
list. This entry will have its coordinate and the branch (unique exit
cell is accepted) joined by a color (:). This means to match in either
templates or controllers, they first must be split.

All remaining references to Chicago, or to the chicago exit cell or
"neighbor", are also removed.

Note this removes validation of split city branches on import, since that
check needs the tile and its orientation, which aren't available at that
point.