Standardize station offsets.
Some of the pre-existing offsets require adjustment to account for the
tile orientation change from earlier. But more importantly, this more
generically handle station placement. It allows rotation to be defined,
to allow for station clusters that don't have a slot at orientation 0.
It also allows a tile to state that its offset is the same as another.
And makes sure to handle station branches.