Setup migration endpoints.
This system allows for the stored data in the original 1846 app to be
transferred over. Since localStorage can only be accessed by a browser
on the page, it requires a few steps: sending the data over, then
visiting the page.

The received data is validated to reduce the attack surface. If it looks
good, a key is generated and it's stored in Redis. The key is then
returned, which must be passed to the migration completion endpoint.

The completion endpoint retrieves the data associated with the key,
again validates it, then stores it in a session cookie. This allows
the browser to redirect to the main 1846 page controller.

The controller loads the data in the session cookie, validates it, then
converts hideCityPaths to hideStopPaths and rotates tiles as needed.

This data is then embedded in the main page template in calls to
localStorage, such that it will be saved before the app starts loading
and displaying the game state.