Handle tile IDs and orientations as ints.
I don't recall how, but the localStorage data for the 1846 Route Finder
can include the tile IDs and orientations as ints. The validation and
conversion functions of the migrate feature need to account for that,
which they can do by simply converting the values to strs before
operating on them.