Highlight tiles with too many stations.
If a tile has too many stations, or there's no slot remaining for a home
or reserved station, it is outlined in red. An error message is displayed
in the tile selector modal across all tabs. To facilitate this, the tile
selection viewport is resized as needed to keep the message on the screen,
and to prevent a double scroll bar.