IN PROGRESS: Update generation of clickable spaces.
This is primarily to make termini with a home station unclickable. It
should also make a built-in gray city with just a home station
unclickable.

I need to mark each space which is marked for a home station upon railroad load, not just for the railroads which are in play. This requires a change in routes-18xx.